requirements.txt
astroid==2.2.4
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-phonenumber-field==2.2.0
idna==2.8
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
phonenumbers==8.10.6
psycopg2-binary==2.7.7
pylint==2.3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
